networksecurity/requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart
pymongo[srv]==3.12